+++ packages/backend-proxy-middleware/test/ext/bsp.test.ts
@@ -1,6 +1,6 @@
 //import prompts from 'prompts';
 import { ToolsLogger, NullTransport } from '@sap-ux/logger';
-import { addOptionsForEmbeddedBSP, convertAppDescriptorToManifest, promptUserPass } from '../../src/ext/bsp';
+import { addOptionsForEmbeddedBSP, promptUserPass } from '../../src/ext/bsp';
 
 // mock required btp-utils functions
 import { isAppStudio } from '@sap-ux/btp-utils';
@@ -22,14 +22,6 @@ describe('bsp', () => {
 
     beforeEach(() => {
         jest.clearAllMocks();
-    });
-
-    test('convertAppDescriptorToManifest', () => {
-        const rewrite = convertAppDescriptorToManifest('/my/bsp');
-        expect(rewrite('/my/bsp/manifest.appdescr')).toBe('/manifest.json');
-        expect(rewrite('/another/manifest.appdescr')).toBe('/another/manifest.appdescr');
-        expect(rewrite('/my/bsp/test')).toBe('/my/bsp/test');
-        expect(rewrite('/test')).toBe('/test');
     });
 
     describe('promptUserPass', () => {
@@ -59,8 +51,7 @@ describe('bsp', () => {
 
     describe('addOptionsForEmbeddedBSP', () => {
         test('standard scenario', async () => {
-            const oldPathRewrite = jest.fn().mockReturnValue('/mocked');
-            const options: Options = { pathRewrite: oldPathRewrite };
+            const options: Options = {};
             await addOptionsForEmbeddedBSP('/my/bsp', options, logger);
 
             expect(options.router).toBeDefined();
@@ -75,11 +66,6 @@ describe('bsp', () => {
             ).toBe('http://local.example');
 
             expect(options.auth).toBeDefined();
-
-            expect(options.pathRewrite).toBeDefined();
-            expect(options.pathRewrite).not.toBe(oldPathRewrite);
-            (options.pathRewrite as Function)('test');
-            expect(oldPathRewrite).toHaveBeenCalled();
         });
 
         test('no existing options', async () => {
@@ -87,7 +73,6 @@ describe('bsp', () => {
             await addOptionsForEmbeddedBSP('/my/bsp', options, logger);
 
             expect(options.router).toBeDefined();
-            expect(options.pathRewrite).toBeDefined();
             expect(options.auth).toBeDefined();
         });
 
@@ -96,7 +81,6 @@ describe('bsp', () => {
             await addOptionsForEmbeddedBSP('/my/bSp', options, logger);
 
             expect(options.router).toBeDefined();
-            expect(options.pathRewrite).toBeDefined();
             expect(options.auth).toBeDefined();
         });
     });