+++ packages/create/src/cli/add/cds-plugin-ui.ts
@@ -1,4 +1,4 @@
-import { relative } from 'path';
+import { relative } from 'node:path';
 import type { Command } from 'commander';
 import { enableCdsUi5Plugin } from '@sap-ux/cap-config-writer';
 import { getLogger, setLogLevelVerbose, traceChanges } from '../../tracing';