+++ packages/fe-fpm-writer/src/filter/index.ts
@@ -2,7 +2,7 @@ import { create as createStorage } from 'mem-fs';
 import type { Editor } from 'mem-fs-editor';
 import { create } from 'mem-fs-editor';
 import type { CustomFilter, InternalCustomFilter, PageOptions } from './types';
-import { join } from 'path';
+import { join } from 'node:path';
 import { render } from 'ejs';
 import { validateBasePath } from '../common/validate';
 import type { Manifest } from '../common/types';