+++ packages/fiori-generator-shared/src/logging/logWrapper.ts
@@ -132,7 +132,7 @@ export class LogWrapper implements ILogWrapper, SapUxLogger {
             }
             LogWrapper._logLevel = logLevel === 'off' || !logLevel ? 'info' : logLevel;
         }
-        LogWrapper._vscodeLogger?.debug(t('debug.loggingConfigured', { logLevel: LogWrapper._logLevel }));
+        LogWrapper._vscodeLogger?.debug(t('logMessages.debug.loggingConfigured', { logLevel: LogWrapper._logLevel }));
     }
 
     static readonly logAtLevel = (level: LogLevel, message: string | object, ...args: any[]) => {
@@ -156,7 +156,7 @@ export class LogWrapper implements ILogWrapper, SapUxLogger {
                 );
             }
         } else {
-            DefaultLogger.error(t('error.logWrapperNotInitialised'));
+            DefaultLogger.error(t('logMessages.error.logWrapperNotInitialised'));
         }
     };