+++ packages/ui-prompting/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sap-ux/ui-prompting
 
+## 0.5.3
+
+### Patch Changes
+
+-   Updated dependencies [4c448a7]
+    -   @sap-ux/ui-components@1.26.12
+
+## 0.5.2
+
+### Patch Changes
+
+-   43a2446: chore: fix Sonar issues
+
 ## 0.5.1
 
 ### Patch Changes